requirements.txt
python-telegram-bot

google-api-python-client
oauth2client
flask
pycryptodome
gunicorn
psycopg2-binary
html2text
python-dateutil